Refactor MoE benchmark and publish workflow
