+++ mesh-llm/src/system/moe_planner.rs
@@ -1,3 +1,7 @@
+mod catalog_resolution;
+mod fit;
+mod package;
+
 use anyhow::{bail, Context, Result};
 use hf_hub::{RepoDownloadFileParams, RepoInfoParams};
 use serde::{Deserialize, Serialize};
@@ -13,6 +17,28 @@ use crate::inference::{election, moe};
 use crate::models::{
     build_hf_api, catalog, gguf::GgufTensorByteProfile, huggingface_identity_for_path,
     resolve_model_spec, resolve_model_spec_with_progress,
+};
+
+#[cfg(test)]
+use self::catalog_resolution::{
+    analyzer_id_from_analysis, join_repo_relative, select_preferred_ranking,
+};
+pub(crate) use self::catalog_resolution::{
+    catalog_entry_path_for_source_repo, fetch_remote_package_expert_components,
+    resolve_runtime_ranking, sort_catalog_package_pointers,
+};
+use self::catalog_resolution::{fetch_remote_package_ranking, resolve_best_ranking};
+use self::fit::{
+    aggregate_tensor_byte_profile, build_mass_checkpoints, expert_bytes_json,
+    predict_plan_fit_for_nodes,
+};
+pub(crate) use self::fit::{estimate_startup_fit_from_analysis, fetch_remote_startup_fit};
+pub(crate) use self::package::{
+    analysis_log_path, build_meshllm_descriptor, build_submit_bundle,
+    canonical_model_ref_from_source, default_package_repo_name_for_model,
+    infer_variant_key_from_gguf_file, normalize_variant_selector, read_analysis_json, sha256_file,
+    variant_key_for_model, write_analysis_json, write_package_analysis_json,
+    write_package_benchmark_json,
 };
 
 const DEFAULT_DATASET_REVISION: &str = "main";
@@ -157,6 +183,18 @@ pub(crate) enum MoeAnalysisExpertBytes {
     DensePerExpert { values: Vec<u64> },
 }
 
+impl MoeAnalysisExpertBytes {
+    fn bytes_for(&self, expert_id: u32) -> Result<u64> {
+        match self {
+            Self::Uniform { bytes_per_expert } => Ok(*bytes_per_expert),
+            Self::DensePerExpert { values } => values
+                .get(expert_id as usize)
+                .copied()
+                .ok_or_else(|| anyhow::anyhow!("expert id {} missing from analysis", expert_id)),
+        }
+    }
+}
+
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub(crate) struct MoeStartupFitEstimate {
     pub analyzer_id: String,
@@ -294,22 +332,22 @@ pub(crate) async fn plan_moe(args: MoePlanArgs) -> Result<MoePlanReport> {
                     required_experts_per_node,
                 )?;
             assumptions.push(format!(
-                    "Estimated shard launch for {} nodes from analysis.json: {:.1}GB max per node ({:.1}GB raw shard bytes)",
-                    recommended_nodes,
-                    predicted_max_shard_launch_bytes as f64 / 1e9,
-                    predicted_max_shard_bytes as f64 / 1e9
-                ));
+                "Estimated shard launch for {} nodes from analysis.json: {:.1}GB max per node ({:.1}GB raw shard bytes)",
+                recommended_nodes,
+                predicted_max_shard_launch_bytes as f64 / 1e9,
+                predicted_max_shard_bytes as f64 / 1e9
+            ));
         }
         if args.nodes.is_none() && !fit.any_fit_exists() {
             assumptions.push(
-                    "No node count up to the current max useful node count fits this VRAM target under the conservative shared-core heuristic."
-                        .to_string(),
-                );
+                "No node count up to the current max useful node count fits this VRAM target under the conservative shared-core heuristic."
+                    .to_string(),
+            );
         } else if args.nodes.is_some() && !feasible {
             assumptions.push(format!(
-                    "Requested node count {} does not fit this VRAM target under the current shared-core heuristic.",
-                    recommended_nodes
-                ));
+                "Requested node count {} does not fit this VRAM target under the current shared-core heuristic.",
+                recommended_nodes
+            ));
         }
 
         (
@@ -475,11 +513,11 @@ pub(crate) fn local_submit_ranking(model: &MoeModelContext) -> Result<ResolvedRa
     let path = moe::package_cache_ranking_path(&model.path);
     if !path.exists() {
         bail!(
-                "No local ranking artifact found for {}. Run `mesh-llm moe analyze full {}` or `mesh-llm moe analyze micro {}` first.",
-                model.display_name,
-                model.input,
-                model.input
-            );
+            "No local ranking artifact found for {}. Run `mesh-llm moe analyze full {}` or `mesh-llm moe analyze micro {}` first.",
+            model.display_name,
+            model.input,
+            model.input
+        );
     };
     Ok(ResolvedRanking {
         analysis_path: sibling_analysis_path(&path),
@@ -556,1230 +594,6 @@ fn sibling_analysis_path(path: &Path) -> Option<PathBuf> {
     let parent = path.parent()?;
     let analysis = parent.join(ANALYSIS_JSON_FILENAME);
     analysis.exists().then_some(analysis)
-}
-
-pub(crate) fn resolve_runtime_ranking(
-    model_path: &Path,
-    catalog_repo_name: &str,
-) -> Result<Option<ResolvedRanking>> {
-    let local_legacy = resolve_local_runtime_ranking(model_path);
-
-    let Some(identity) = huggingface_identity_for_path(model_path) else {
-        return Ok(local_legacy);
-    };
-
-    if local_legacy
-        .as_ref()
-        .is_some_and(|ranking| ranking_method_priority(ranking) >= 2)
-    {
-        // A local full-v1 ranking is treated as the current best tie-break because
-        // we do not have a comparable freshness signal for local cache entries.
-        return Ok(local_legacy);
-    }
-
-    let model_ref = canonical_model_ref_from_source(
-        &identity.repo_id,
-        &normalize_distribution_id(&identity.local_file_name),
-        Some(identity.local_file_name.clone()),
-    );
-    let remote = match fetch_remote_package_ranking(catalog_repo_name, &model_ref, false) {
-        Ok(remote) => remote,
-        Err(error) => return local_legacy.ok_or(error).map(Some),
-    };
-    Ok(select_preferred_ranking(local_legacy, remote))
-}
-
-type AnalysisDetails = (Option<&'static str>, Option<usize>, u32, u32, bool);
-
-fn resolve_local_runtime_ranking(model_path: &Path) -> Option<ResolvedRanking> {
-    let path = moe::package_cache_ranking_path(model_path);
-    if !path.exists() {
-        return None;
-    }
-    let analysis_path = moe::package_cache_analysis_path(model_path);
-    let analyzer_id =
-        analyzer_id_from_analysis(&analysis_path).unwrap_or_else(|_| "full-v1".to_string());
-    Some(ResolvedRanking {
-        path,
-        metadata_path: None,
-        analysis_path: analysis_path.exists().then_some(analysis_path),
-        analyzer_id,
-        source: RankingSource::LocalCache,
-        reason: "local materialized package ranking".to_string(),
-    })
-}
-
-async fn resolve_best_ranking(
-    model: &MoeModelContext,
-    args: &MoePlanArgs,
-) -> Result<ResolvedRanking> {
-    let local_legacy = resolve_local_runtime_ranking(&model.path);
-
-    let Some(source_repo) = model.source_repo.as_ref() else {
-        return local_legacy.ok_or_else(|| {
-            anyhow::anyhow!(
-                "No published ranking lookup is possible for non-HF model {} and no local ranking cache exists.",
-                model.display_name
-            )
-        });
-    };
-    if local_legacy
-        .as_ref()
-        .is_some_and(|ranking| ranking_method_priority(ranking) >= 2)
-    {
-        // Prefer local full-v1 on tie: remote is only selected when it is stronger.
-        return Ok(local_legacy.expect("checked is_some above"));
-    }
-
-    let model_ref = canonical_model_ref_from_source(
-        source_repo,
-        &model.distribution_id,
-        Some(model.distribution_id.clone()),
-    );
-    let local_fallback = local_legacy.clone();
-    let mut remote_spinner = args.progress.then(|| {
-        start_spinner(&format!(
-            "Checking published MoE package for {}",
-            model.display_name
-        ))
-    });
-    if let Some(spinner) = remote_spinner.as_mut() {
-        spinner.set_message(format!(
-            "Fetching published MoE package ranking and analysis for {}",
-            model.display_name
-        ));
-    }
-    let remote_lookup = fetch_remote_package_ranking(&args.catalog_repo, &model_ref, args.progress);
-    if let Some(spinner) = remote_spinner.as_mut() {
-        spinner.finish();
-    }
-
-    let remote = match remote_lookup {
-        Ok(remote) => remote,
-        Err(error) => {
-            return local_fallback.ok_or(error);
-        }
-    };
-    select_preferred_ranking(local_legacy, remote).ok_or_else(|| {
-        anyhow::anyhow!(
-            "No published MoE package ranking found in {} for {}:{} and no local cache exists.",
-            args.catalog_repo,
-            source_repo,
-            model.distribution_id
-        )
-    })
-}
-
-fn ranking_method_priority(ranking: &ResolvedRanking) -> u8 {
-    if ranking.analyzer_id.starts_with("full") {
-        2
-    } else if ranking.analyzer_id.starts_with("micro") {
-        1
-    } else {
-        0
-    }
-}
-
-fn select_preferred_ranking(
-    local: Option<ResolvedRanking>,
-    remote: Option<ResolvedRanking>,
-) -> Option<ResolvedRanking> {
-    match (local, remote) {
-        (Some(local), Some(remote)) => {
-            if ranking_method_priority(&local) >= ranking_method_priority(&remote) {
-                Some(local)
-            } else {
-                Some(remote)
-            }
-        }
-        (Some(local), None) => Some(local),
-        (None, Some(remote)) => Some(remote),
-        (None, None) => None,
-    }
-}
-
-pub(crate) fn catalog_entry_path_for_source_repo(source_repo: &str) -> Result<String> {
-    let (source_owner, source_name) = source_repo
-        .split_once('/')
-        .ok_or_else(|| anyhow::anyhow!("Invalid source repo {}", source_repo))?;
-    anyhow::ensure!(
-        !source_owner.is_empty()
-            && !source_name.is_empty()
-            && source_owner != "."
-            && source_owner != ".."
-            && source_name != "."
-            && source_name != ".."
-            && !source_name.contains('/'),
-        "Invalid source repo {}",
-        source_repo
-    );
-    Ok(format!("entries/{source_owner}/{source_name}.json",))
-}
-
-fn package_variant_for_model_ref(model_ref: &str) -> Result<&str> {
-    model_ref
-        .rsplit_once(':')
-        .map(|(_, variant)| variant)
-        .ok_or_else(|| anyhow::anyhow!("Invalid Mesh model ref {model_ref}"))
-}
-
-fn join_repo_relative(parent_file: &str, relative_path: &str) -> String {
-    let relative = relative_path.trim_start_matches('/');
-    let Some(parent) = Path::new(parent_file).parent() else {
-        return relative.to_string();
-    };
-    let joined = parent.join(relative);
-    joined.to_string_lossy().replace('\\', "/")
-}
-
-fn analyzer_id_from_analysis(analysis_path: &Path) -> Result<String> {
-    let analysis_content = fs::read_to_string(analysis_path)
-        .with_context(|| format!("Read {}", analysis_path.display()))?;
-    let parsed: Value = serde_json::from_str(&analysis_content)
-        .with_context(|| format!("Parse {}", analysis_path.display()))?;
-    Ok(parsed
-        .get("analyzer_id")
-        .and_then(Value::as_str)
-        .unwrap_or("full-v1")
-        .to_string())
-}
-
-fn resolve_catalog_package_pointer(
-    api: &hf_hub::HFClientSync,
-    catalog_repo_name: &str,
-    model_ref: &str,
-) -> Result<Option<crate::models::catalog::CatalogPackagePointer>> {
-    let (owner, name) = catalog_repo_name
-        .split_once('/')
-        .unwrap_or(("", catalog_repo_name));
-    let catalog_repo = api.dataset(owner, name);
-    let info = catalog_repo.info(
-        &RepoInfoParams::builder()
-            .revision(DEFAULT_DATASET_REVISION.to_string())
-            .build(),
-    )?;
-    let hf_hub::RepoInfo::Dataset(info) = info else {
-        bail!("Expected dataset repo info for {}", catalog_repo_name);
-    };
-
-    let dataset_revision = info.sha.as_deref().unwrap_or(DEFAULT_DATASET_REVISION);
-    let (source_repo, variant) = model_ref
-        .rsplit_once(':')
-        .ok_or_else(|| anyhow::anyhow!("Invalid Mesh model ref {model_ref}"))?;
-    let entry_path = catalog_entry_path_for_source_repo(source_repo)?;
-    if !info
-        .siblings
-        .as_deref()
-        .unwrap_or(&[])
-        .iter()
-        .any(|entry| entry.rfilename == entry_path)
-    {
-        return Ok(None);
-    }
-
-    let downloaded = catalog_repo
-        .download_file(
-            &RepoDownloadFileParams::builder()
-                .filename(entry_path.clone())
-                .revision(dataset_revision.to_string())
-                .build(),
-        )
-        .with_context(|| format!("Download {}", entry_path))?;
-    let content = fs::read_to_string(&downloaded)
-        .with_context(|| format!("Read {}", downloaded.display()))?;
-    let entry: crate::models::catalog::CatalogRepoEntry = serde_json::from_str(&content)
-        .with_context(|| format!("Parse {}", downloaded.display()))?;
-    let mut packages = entry
-        .variants
-        .get(variant)
-        .map(|variant_entry| variant_entry.packages.clone())
-        .unwrap_or_default();
-    sort_catalog_package_pointers(&mut packages);
-    Ok(packages.into_iter().next())
-}
-
-pub(crate) fn sort_catalog_package_pointers(
-    packages: &mut [crate::models::catalog::CatalogPackagePointer],
-) {
-    packages.sort_by(compare_catalog_package_pointers);
-}
-
-fn compare_catalog_package_pointers(
-    left: &crate::models::catalog::CatalogPackagePointer,
-    right: &crate::models::catalog::CatalogPackagePointer,
-) -> std::cmp::Ordering {
-    catalog_package_trust_rank(&right.trust)
-        .cmp(&catalog_package_trust_rank(&left.trust))
-        .then_with(|| left.publisher.cmp(&right.publisher))
-        .then_with(|| left.package_repo.cmp(&right.package_repo))
-        .then_with(|| left.package_revision.cmp(&right.package_revision))
-}
-
-fn catalog_package_trust_rank(trust: &str) -> u8 {
-    match trust {
-        "canonical" => 2,
-        "community" => 1,
-        _ => 0,
-    }
-}
-
-fn resolve_package_variant(
-    api: &hf_hub::HFClientSync,
-    package_repo: &str,
-    package_revision: &str,
-    model_ref: &str,
-) -> Result<Option<(Vec<hf_hub::RepoSibling>, String, String)>> {
-    let (owner, name) = package_repo.split_once('/').unwrap_or(("", package_repo));
-    let repo = api.model(owner, name);
-    let info = repo.info(
-        &RepoInfoParams::builder()
-            .revision(package_revision.to_string())
-            .build(),
-    )?;
-    let hf_hub::RepoInfo::Model(info) = info else {
-        bail!("Expected model repo info for {}", package_repo);
-    };
-    let resolved_revision = info.sha.unwrap_or_else(|| package_revision.to_string());
-    let siblings = info.siblings.unwrap_or_default();
-    if !siblings
-        .iter()
-        .any(|entry| entry.rfilename == "meshllm.json")
-    {
-        return Ok(None);
-    }
-
-    let meshllm_path = repo
-        .download_file(
-            &RepoDownloadFileParams::builder()
-                .filename("meshllm.json".to_string())
-                .revision(resolved_revision.clone())
-                .build(),
-        )
-        .with_context(|| format!("Download meshllm.json from {}", package_repo))?;
-    let meshllm_content = fs::read_to_string(&meshllm_path)
-        .with_context(|| format!("Read {}", meshllm_path.display()))?;
-    let meshllm: MeshllmPackageJson = serde_json::from_str(&meshllm_content)
-        .with_context(|| format!("Parse {}", meshllm_path.display()))?;
-    let variant = package_variant_for_model_ref(model_ref)?;
-    let Some(variant_entry) = meshllm.variants.get(variant) else {
-        return Ok(None);
-    };
-    Ok(Some((
-        siblings,
-        resolved_revision,
-        variant_entry.manifest.clone(),
-    )))
-}
-
-fn fetch_remote_package_ranking(
-    catalog_repo_name: &str,
-    model_ref: &str,
-    progress: bool,
-) -> Result<Option<ResolvedRanking>> {
-    let catalog_repo_name = catalog_repo_name.to_string();
-    let model_ref = model_ref.to_string();
-    crate::models::run_hf_blocking(move || {
-        let api =
-            build_hf_api(progress).context("Build Hugging Face client for MoE package lookup")?;
-        let Some(entry) = resolve_catalog_package_pointer(&api, &catalog_repo_name, &model_ref)?
-        else {
-            return Ok(None);
-        };
-        let Some((siblings, package_revision, manifest_path)) = resolve_package_variant(
-            &api,
-            &entry.package_repo,
-            &entry.package_revision,
-            &model_ref,
-        )?
-        else {
-            return Ok(None);
-        };
-
-        let ranking_rel = join_repo_relative(&manifest_path, "ranking.csv");
-        let analysis_rel = join_repo_relative(&manifest_path, ANALYSIS_JSON_FILENAME);
-        if !(siblings.iter().any(|entry| entry.rfilename == ranking_rel)
-            && siblings.iter().any(|entry| entry.rfilename == analysis_rel))
-        {
-            return Ok(None);
-        }
-
-        let (owner, name) = entry
-            .package_repo
-            .split_once('/')
-            .unwrap_or(("", entry.package_repo.as_str()));
-        let repo = api.model(owner, name);
-        let ranking_path = repo
-            .download_file(
-                &RepoDownloadFileParams::builder()
-                    .filename(ranking_rel.clone())
-                    .revision(package_revision.clone())
-                    .build(),
-            )
-            .with_context(|| format!("Download {}", ranking_rel))?;
-        let analysis_path = repo
-            .download_file(
-                &RepoDownloadFileParams::builder()
-                    .filename(analysis_rel.clone())
-                    .revision(package_revision.clone())
-                    .build(),
-            )
-            .with_context(|| format!("Download {}", analysis_rel))?;
-        let analyzer_id = analyzer_id_from_analysis(&analysis_path)?;
-        Ok(Some(ResolvedRanking {
-            path: ranking_path,
-            metadata_path: None,
-            analysis_path: Some(analysis_path),
-            analyzer_id,
-            source: RankingSource::HuggingFacePackage,
-            reason: format!("published package ranking in {}", entry.package_repo),
-        }))
-    })
-}
-
-pub(crate) fn fetch_remote_package_expert_components(
-    catalog_repo_name: &str,
-    model_ref: &str,
-    ranking_sha256: &str,
-    expected_experts: &[u32],
-    progress: bool,
-) -> Result<Option<ResolvedExpertComponents>> {
-    let catalog_repo_name = catalog_repo_name.to_string();
-    let model_ref = model_ref.to_string();
-    let ranking_sha256 = ranking_sha256.to_string();
-    let expected_experts = expected_experts.to_vec();
-    crate::models::run_hf_blocking(move || {
-        let api = build_hf_api(progress)
-            .context("Build Hugging Face client for MoE expert component lookup")?;
-        let Some(entry) = resolve_catalog_package_pointer(&api, &catalog_repo_name, &model_ref)?
-        else {
-            return Ok(None);
-        };
-        let Some((siblings, package_revision, manifest_rel)) = resolve_package_variant(
-            &api,
-            &entry.package_repo,
-            &entry.package_revision,
-            &model_ref,
-        )?
-        else {
-            return Ok(None);
-        };
-        if !siblings
-            .iter()
-            .any(|sibling| sibling.rfilename == manifest_rel)
-        {
-            return Ok(None);
-        }
-
-        let (owner, name) = entry
-            .package_repo
-            .split_once('/')
-            .unwrap_or(("", entry.package_repo.as_str()));
-        let repo = api.model(owner, name);
-        let manifest_path = repo
-            .download_file(
-                &RepoDownloadFileParams::builder()
-                    .filename(manifest_rel.clone())
-                    .revision(package_revision.clone())
-                    .build(),
-            )
-            .with_context(|| format!("Download {}", manifest_rel))?;
-        let manifest_content = fs::read_to_string(&manifest_path)
-            .with_context(|| format!("Read {}", manifest_path.display()))?;
-        let manifest: MoePackageManifest = serde_json::from_str(&manifest_content)
-            .with_context(|| format!("Parse {}", manifest_path.display()))?;
-
-        if manifest.ranking_sha256 != ranking_sha256 || manifest.format != "meshllm-moe-components"
-        {
-            return Ok(None);
-        }
-
-        let trunk_rel = join_repo_relative(&manifest_rel, &manifest.trunk.path);
-        if !siblings.iter().any(|entry| entry.rfilename == trunk_rel) {
-            return Ok(None);
-        }
-        let trunk_path = repo
-            .download_file(
-                &RepoDownloadFileParams::builder()
-                    .filename(trunk_rel.clone())
-                    .revision(package_revision.clone())
-                    .build(),
-            )
-            .with_context(|| format!("Download {}", trunk_rel))?;
-        if sha256_file(&trunk_path)? != manifest.trunk.sha256 {
-            bail!("Downloaded trunk component hash mismatch for {}", trunk_rel);
-        }
-
-        let mut expert_paths = Vec::with_capacity(expected_experts.len());
-        for expert_id in &expected_experts {
-            let Some(expert) = manifest
-                .experts
-                .iter()
-                .find(|entry| entry.expert_id == Some(*expert_id))
-            else {
-                return Ok(None);
-            };
-            let expert_rel = join_repo_relative(&manifest_rel, &expert.path);
-            if !siblings.iter().any(|entry| entry.rfilename == expert_rel) {
-                return Ok(None);
-            }
-            let expert_path = repo
-                .download_file(
-                    &RepoDownloadFileParams::builder()
-                        .filename(expert_rel.clone())
-                        .revision(package_revision.clone())
-                        .build(),
-                )
-                .with_context(|| format!("Download {}", expert_rel))?;
-            if sha256_file(&expert_path)? != expert.sha256 {
-                bail!(
-                    "Downloaded expert component hash mismatch for {}",
-                    expert_rel
-                );
-            }
-            expert_paths.push(expert_path);
-        }
-
-        Ok(Some(ResolvedExpertComponents {
-            prefix: format!(
-                "{}@{}:{}",
-                entry.package_repo, package_revision, manifest_rel
-            ),
-            trunk_path,
-            expert_paths,
-        }))
-    })
-}
-
-pub(crate) fn build_submit_bundle(
-    model: &MoeModelContext,
-    ranking: &ResolvedRanking,
-    log_path: Option<&Path>,
-) -> Result<MoeSubmitBundle> {
-    let variant = variant_key_for_model(model);
-    let variant_root = format!("variants/{variant}");
-    let ranking_rel = format!("{variant_root}/ranking.csv");
-    let analysis_rel = format!("{variant_root}/{ANALYSIS_JSON_FILENAME}");
-    let manifest_rel = format!("{variant_root}/manifest.json");
-    let log_rel = format!("{variant_root}/run.log");
-
-    let log_path = log_path.filter(|path| path.exists()).map(Path::to_path_buf);
-    let model_ref = canonical_model_ref_for_model(model)?;
-    Ok(MoeSubmitBundle {
-        model_ref,
-        variant,
-        variant_root,
-        ranking_repo_path: ranking_rel,
-        analysis_repo_path: analysis_rel,
-        manifest_repo_path: manifest_rel,
-        log_repo_path: log_path.as_ref().map(|_| log_rel),
-        commit_message: format!(
-            "Publish {} {} package",
-            model.distribution_id, ranking.analyzer_id,
-        ),
-        commit_description: format!(
-            "Publish {} Mesh package artifacts for {} ({})",
-            ranking.analyzer_id, model.display_name, model.input
-        ),
-    })
-}
-
-pub(crate) fn build_meshllm_descriptor(
-    existing: Option<&str>,
-    model: &MoeModelContext,
-    manifest_path: &str,
-) -> Result<MeshllmPackageJson> {
-    let source_repo = model.source_repo.clone().ok_or_else(|| {
-        anyhow::anyhow!("Package publishing requires a Hugging Face-backed model")
-    })?;
-    let source_revision = model.source_revision.clone().ok_or_else(|| {
-        anyhow::anyhow!("Package publishing requires a resolved Hugging Face source revision")
-    })?;
-    let variant = variant_key_for_model(model);
-    let mut meshllm = if let Some(existing) = existing {
-        let parsed: MeshllmPackageJson =
-            serde_json::from_str(existing).context("Parse existing meshllm.json")?;
-        if parsed.source.repo != source_repo {
-            bail!(
-                "Existing meshllm.json source repo {} does not match {}",
-                parsed.source.repo,
-                source_repo
-            );
-        }
-        parsed
-    } else {
-        MeshllmPackageJson {
-            schema_version: 1,
-            source: MeshllmPackageSource {
-                repo: source_repo.clone(),
-                revision: source_revision.clone(),
-            },
-            variants: BTreeMap::new(),
-        }
-    };
-    meshllm.source.revision = source_revision;
-    meshllm.variants.insert(
-        variant,
-        MeshllmPackageVariant {
-            distribution_id: model.distribution_id.clone(),
-            manifest: manifest_path.to_string(),
-        },
-    );
-    Ok(meshllm)
-}
-
-fn build_package_analysis_json(
-    model: &MoeModelContext,
-    ranking: &ResolvedRanking,
-    model_files: &[(String, PathBuf)],
-) -> Result<Value> {
-    let (prompt_set, prompt_count, token_count, context_size, all_layers) =
-        infer_analysis_details(ranking, model)?;
-    let mass_profile = load_analyze_mass_profile(&ranking.path)?;
-    let tensor_profile = aggregate_tensor_byte_profile(model_files)?;
-
-    Ok(json!({
-        "schema_version": 1,
-        "analyzer_id": ranking.analyzer_id,
-        "created_at": iso8601_now(),
-        "tool": {
-            "name": "llama-moe-analyze",
-            "version": "mesh-llm-fork"
-        },
-        "parameters": {
-            "prompt_set": prompt_set,
-            "prompt_count": prompt_count,
-            "token_count": token_count,
-            "context_size": context_size,
-            "all_layers": all_layers
-        },
-        "ranking": {
-            "sha256": sha256_file(&ranking.path)?,
-            "rows": mass_profile.ranking.len(),
-            "mass_checkpoints": build_mass_checkpoints(&mass_profile),
-        },
-        "summary": {
-            "n_expert": model.expert_count,
-            "n_expert_used": model.used_expert_count,
-            "min_experts_per_node": model.min_experts_per_node
-        },
-        "memory": {
-            "full_model_bytes": tensor_profile.full_model_bytes,
-            "base_resident_bytes": tensor_profile.base_resident_bytes,
-            "shard_file_overhead_bytes": tensor_profile.file_overhead_bytes,
-            "expert_tensor_bytes_total": tensor_profile.expert_tensor_bytes,
-            "expert_bytes": expert_bytes_json(tensor_profile.expert_tensor_bytes, model.expert_count)?,
-        },
-        "planner": {
-            "recommended_overlap": 1,
-        },
-        "artifacts": {
-            "ranking": "ranking.csv",
-            "manifest": "manifest.json"
-        }
-    }))
-}
-
-fn build_analysis_json(
-    model: &MoeModelContext,
-    ranking: &ResolvedRanking,
-    model_files: &[(String, PathBuf)],
-) -> Result<Value> {
-    let mass_profile = load_analyze_mass_profile(&ranking.path)?;
-    let tensor_profile = aggregate_tensor_byte_profile(model_files)?;
-    Ok(json!({
-        "schema_version": 1,
-        "ranking": {
-            "sha256": sha256_file(&ranking.path)?,
-            "rows": mass_profile.ranking.len(),
-            "mass_checkpoints": build_mass_checkpoints(&mass_profile),
-        },
-        "model": {
-            "expert_count": model.expert_count,
-            "expert_used_count": model.used_expert_count,
-        },
-        "memory": {
-            "full_model_bytes": tensor_profile.full_model_bytes,
-            "base_resident_bytes": tensor_profile.base_resident_bytes,
-            "shard_file_overhead_bytes": tensor_profile.file_overhead_bytes,
-            "expert_tensor_bytes_total": tensor_profile.expert_tensor_bytes,
-            "expert_bytes": expert_bytes_json(tensor_profile.expert_tensor_bytes, model.expert_count)?,
-        }
-    }))
-}
-
-pub(crate) fn default_package_repo_name_for_model(model: &MoeModelContext) -> Result<String> {
-    let Some(source_repo) = model.source_repo.as_ref() else {
-        bail!("A Hugging Face-backed model is required to derive the package repo name.");
-    };
-    let repo_name = source_repo
-        .split_once('/')
-        .map(|(_, repo)| repo)
-        .unwrap_or(source_repo);
-    Ok(format!("{}-moe", sanitize_repo_name_component(repo_name)))
-}
-
-pub(crate) fn canonical_model_ref_for_model(model: &MoeModelContext) -> Result<String> {
-    let Some(source_repo) = model.source_repo.as_ref() else {
-        bail!("A Hugging Face-backed model is required to derive the canonical model ref.");
-    };
-    Ok(canonical_model_ref_from_source(
-        source_repo,
-        &model.distribution_id,
-        Some(
-            crate::models::huggingface_identity_for_path(&model.path)
-                .map(|identity| identity.file)
-                .or_else(|| {
-                    model
-                        .path
-                        .file_name()
-                        .and_then(|value| value.to_str())
-                        .map(ToOwned::to_owned)
-                })
-                .unwrap_or_else(|| model.display_name.clone()),
-        ),
-    ))
-}
-
-pub(crate) fn canonical_model_ref_from_source(
-    source_repo: &str,
-    distribution_id: &str,
-    file_hint: Option<String>,
-) -> String {
-    let variant = file_hint
-        .as_deref()
-        .and_then(infer_variant_key_from_gguf_file)
-        .unwrap_or_else(|| normalize_variant_selector(distribution_id));
-    format!("{source_repo}:{variant}")
-}
-
-pub(crate) fn variant_key_for_model(model: &MoeModelContext) -> String {
-    let candidate = crate::models::huggingface_identity_for_path(&model.path)
-        .map(|identity| identity.file)
-        .or_else(|| {
-            model
-                .path
-                .file_name()
-                .and_then(|value| value.to_str())
-                .map(ToOwned::to_owned)
-        })
-        .unwrap_or_else(|| model.display_name.clone());
-    infer_variant_key_from_gguf_file(&candidate).unwrap_or_else(|| model.distribution_id.clone())
-}
-
-pub(crate) fn sanitize_repo_name_component(input: &str) -> String {
-    let sanitized = input
-        .chars()
-        .map(|ch| {
-            if ch.is_ascii_alphanumeric() || matches!(ch, '-' | '_' | '.') {
-                ch.to_ascii_lowercase()
-            } else {
-                '-'
-            }
-        })
-        .collect::<String>();
-    let trimmed = sanitized.trim_matches('-');
-    if trimmed.is_empty() {
-        "mesh-llm-moe".to_string()
-    } else {
-        trimmed.to_string()
-    }
-}
-
-pub(crate) fn infer_variant_key_from_gguf_file(file: &str) -> Option<String> {
-    if !file.ends_with(".gguf") {
-        return None;
-    }
-    if let Some((prefix, _)) = file.split_once('/') {
-        if is_quant_like_selector(prefix) {
-            return Some(normalize_variant_selector(prefix));
-        }
-    }
-    let basename = Path::new(file).file_name()?.to_str()?;
-    let mut stem = basename.strip_suffix(".gguf")?;
-    if let Some((base, shard)) = stem.rsplit_once("-00001-of-") {
-        if shard.len() == 5 && shard.chars().all(|ch| ch.is_ascii_digit()) {
-            stem = base;
-        }
-    }
-    for marker in [
-        "-UD-", ".UD-", "-IQ", ".IQ", "-Q", ".Q", "-BF16", ".BF16", "-F16", ".F16", "-F32", ".F32",
-    ] {
-        if let Some(pos) = stem.rfind(marker) {
-            return Some(normalize_variant_selector(&stem[pos + 1..]));
-        }
-    }
-    None
-}
-
-pub(crate) fn normalize_variant_selector(selector: &str) -> String {
-    selector.strip_prefix("UD-").unwrap_or(selector).to_string()
-}
-
-fn is_quant_like_selector(value: &str) -> bool {
-    let upper = value.to_ascii_uppercase();
-    upper.starts_with("UD-")
-        || upper.starts_with("Q")
-        || upper.starts_with("IQ")
-        || upper == "BF16"
-        || upper == "F16"
-        || upper == "F32"
-}
-
-pub(crate) fn read_analysis_json(path: &Path) -> Result<MoeAnalysisJson> {
-    let content =
-        fs::read_to_string(path).with_context(|| format!("Read analysis {}", path.display()))?;
-    let parsed: MoeAnalysisJson = serde_json::from_str(&content)
-        .with_context(|| format!("Parse analysis {}", path.display()))?;
-    if parsed.schema_version != 1 {
-        bail!(
-            "Unsupported analysis schema {} in {}",
-            parsed.schema_version,
-            path.display()
-        );
-    }
-    Ok(parsed)
-}
-
-pub(crate) fn estimate_startup_fit_from_analysis(
-    ranking_path: &Path,
-    analysis: &MoeAnalysisJson,
-    target_vram_bytes: u64,
-    analyzer_id: &str,
-    ranking_source: &'static str,
-) -> Result<MoeStartupFitEstimate> {
-    let ranking = moe::load_cached_ranking(ranking_path)
-        .ok_or_else(|| anyhow::anyhow!("Could not parse ranking {}", ranking_path.display()))?;
-    if analysis.ranking.rows != ranking.len() {
-        bail!(
-            "Ranking row count mismatch for {}: analysis says {}, CSV has {}",
-            ranking_path.display(),
-            analysis.ranking.rows,
-            ranking.len()
-        );
-    }
-
-    let required_experts_per_node = default_required_experts_per_node(analysis.model.expert_count);
-    let full_model_launch_bytes = apply_model_load_headroom(analysis.memory.full_model_bytes);
-    let mut estimate = MoeStartupFitEstimate {
-        analyzer_id: analyzer_id.to_string(),
-        ranking_source,
-        target_vram_bytes,
-        required_experts_per_node,
-        full_model_bytes: analysis.memory.full_model_bytes,
-        full_model_launch_bytes,
-        recommended_nodes: None,
-        predicted_max_shard_bytes: None,
-        predicted_max_shard_launch_bytes: None,
-    };
-
-    if estimate.full_model_fits() {
-        estimate.recommended_nodes = Some(1);
-        return Ok(estimate);
-    }
-
-    let max_supported_nodes =
-        (analysis.model.expert_count / required_experts_per_node.max(1)).max(1) as usize;
-    let mut best_failed_shard: Option<(u64, u64)> = None;
-    for nodes in 2..=max_supported_nodes {
-        let assignments =
-            moe::compute_assignments_with_overlap(&ranking, nodes, required_experts_per_node, 1);
-        let max_shard_bytes = assignments
-            .iter()
-            .map(|assignment| predict_shard_bytes(&analysis.memory, &assignment.experts))
-            .collect::<Result<Vec<_>>>()?
-            .into_iter()
-            .max()
-            .unwrap_or_default();
-        let max_shard_launch_bytes = apply_model_load_headroom(max_shard_bytes);
-        if max_shard_launch_bytes <= target_vram_bytes {
-            estimate.recommended_nodes = Some(nodes);
-            estimate.predicted_max_shard_bytes = Some(max_shard_bytes);
-            estimate.predicted_max_shard_launch_bytes = Some(max_shard_launch_bytes);
-            return Ok(estimate);
-        }
-        best_failed_shard = Some((max_shard_bytes, max_shard_launch_bytes));
-    }
-
-    if let Some((max_shard_bytes, max_shard_launch_bytes)) = best_failed_shard {
-        estimate.predicted_max_shard_bytes = Some(max_shard_bytes);
-        estimate.predicted_max_shard_launch_bytes = Some(max_shard_launch_bytes);
-    } else if max_supported_nodes == 1 {
-        let max_shard_bytes = predict_shard_bytes(&analysis.memory, &ranking)?;
-        estimate.predicted_max_shard_bytes = Some(max_shard_bytes);
-        estimate.predicted_max_shard_launch_bytes =
-            Some(apply_model_load_headroom(max_shard_bytes));
-    }
-
-    Ok(estimate)
-}
-
-pub(crate) fn fetch_remote_startup_fit(
-    catalog_repo_name: &str,
-    source_repo: &str,
-    distribution_id: &str,
-    target_vram_bytes: u64,
-    progress: bool,
-) -> Result<Option<MoeStartupFitEstimate>> {
-    let model_ref = canonical_model_ref_from_source(source_repo, distribution_id, None);
-    let Some(ranking) = fetch_remote_package_ranking(catalog_repo_name, &model_ref, progress)?
-    else {
-        return Ok(None);
-    };
-    let Some(analysis_path) = ranking.analysis_path.as_ref() else {
-        return Ok(None);
-    };
-    let analysis = read_analysis_json(analysis_path)?;
-    let estimate = estimate_startup_fit_from_analysis(
-        &ranking.path,
-        &analysis,
-        target_vram_bytes,
-        &ranking.analyzer_id,
-        ranking.source.label(),
-    )?;
-    Ok(Some(estimate))
-}
-
-fn build_mass_checkpoints(profile: &AnalyzeMassProfile) -> Vec<Value> {
-    let mut checkpoints = DEFAULT_MASS_CHECKPOINTS
-        .into_iter()
-        .filter(|top_n| *top_n <= profile.masses.len())
-        .collect::<Vec<_>>();
-    if !checkpoints.contains(&profile.masses.len()) {
-        checkpoints.push(profile.masses.len());
-    }
-    checkpoints
-        .into_iter()
-        .map(|top_n| {
-            let captured_mass = profile
-                .masses
-                .iter()
-                .take(top_n)
-                .map(|(_, mass)| *mass)
-                .sum::<f64>();
-            json!({
-                "top_n": top_n,
-                "mass_fraction": if profile.total_mass > f64::EPSILON {
-                    captured_mass / profile.total_mass
-                } else {
-                    0.0
-                }
-            })
-        })
-        .collect()
-}
-
-fn aggregate_tensor_byte_profile(
-    model_files: &[(String, PathBuf)],
-) -> Result<AggregatedTensorByteProfile> {
-    let mut aggregated = AggregatedTensorByteProfile::default();
-    for (_, path) in model_files {
-        let profile: GgufTensorByteProfile =
-            crate::models::gguf::scan_gguf_tensor_byte_profile(path).ok_or_else(|| {
-                anyhow::anyhow!("Could not derive GGUF tensor bytes for {}", path.display())
-            })?;
-        aggregated.full_model_bytes = aggregated
-            .full_model_bytes
-            .saturating_add(profile.full_model_bytes);
-        aggregated.base_resident_bytes = aggregated
-            .base_resident_bytes
-            .saturating_add(profile.base_resident_bytes);
-        aggregated.expert_tensor_bytes = aggregated
-            .expert_tensor_bytes
-            .saturating_add(profile.expert_tensor_bytes);
-        aggregated.file_overhead_bytes = aggregated
-            .file_overhead_bytes
-            .saturating_add(profile.file_overhead_bytes);
-    }
-    Ok(aggregated)
-}
-
-fn expert_bytes_json(expert_tensor_bytes: u64, expert_count: u32) -> Result<Value> {
-    if expert_count == 0 {
-        bail!("cannot derive expert bytes for expert_count=0");
-    }
-    let expert_count_u64 = u64::from(expert_count);
-    if expert_tensor_bytes % expert_count_u64 == 0 {
-        return Ok(json!({
-            "kind": "uniform",
-            "bytes_per_expert": expert_tensor_bytes / expert_count_u64,
-        }));
-    }
-
-    let base = expert_tensor_bytes / expert_count_u64;
-    let remainder = (expert_tensor_bytes % expert_count_u64) as usize;
-    let mut values = vec![base; expert_count as usize];
-    for value in values.iter_mut().take(remainder) {
-        *value += 1;
-    }
-    Ok(json!({
-        "kind": "dense_per_expert",
-        "values": values,
-    }))
-}
-
-fn default_required_experts_per_node(expert_count: u32) -> u32 {
-    ((expert_count as f64) * 0.5).ceil() as u32
-}
-
-fn apply_model_load_headroom(bytes: u64) -> u64 {
-    bytes
-        .saturating_mul(MODEL_LOAD_HEADROOM_NUMERATOR)
-        .div_ceil(MODEL_LOAD_HEADROOM_DENOMINATOR)
-}
-
-fn predict_plan_fit_for_nodes(
-    ranking: &[u32],
-    analysis: &MoeAnalysisJson,
-    nodes: usize,
-    required_experts_per_node: u32,
-) -> Result<(Vec<moe::NodeAssignment>, u64, u64)> {
-    let assignments =
-        moe::compute_assignments_with_overlap(ranking, nodes, required_experts_per_node, 1);
-    let max_shard_bytes = assignments
-        .iter()
-        .map(|assignment| predict_shard_bytes(&analysis.memory, &assignment.experts))
-        .collect::<Result<Vec<_>>>()?
-        .into_iter()
-        .max()
-        .unwrap_or_default();
-    Ok((
-        assignments,
-        max_shard_bytes,
-        apply_model_load_headroom(max_shard_bytes),
-    ))
-}
-
-fn predict_shard_bytes(memory: &MoeAnalysisMemory, experts: &[u32]) -> Result<u64> {
-    let mut total = memory
-        .base_resident_bytes
-        .saturating_add(memory.shard_file_overhead_bytes);
-    for expert_id in experts {
-        total = total
-            .checked_add(memory.expert_bytes.bytes_for(*expert_id)?)
-            .ok_or_else(|| anyhow::anyhow!("predicted shard bytes overflow"))?;
-    }
-    Ok(total)
-}
-
-impl MoeAnalysisExpertBytes {
-    fn bytes_for(&self, expert_id: u32) -> Result<u64> {
-        match self {
-            Self::Uniform { bytes_per_expert } => Ok(*bytes_per_expert),
-            Self::DensePerExpert { values } => values
-                .get(expert_id as usize)
-                .copied()
-                .ok_or_else(|| anyhow::anyhow!("expert id {} missing from analysis", expert_id)),
-        }
-    }
-}
-
-pub(crate) fn write_analysis_json(
-    model: &MoeModelContext,
-    ranking_path: &Path,
-    analyzer_id: &str,
-) -> Result<PathBuf> {
-    let ranking = ResolvedRanking {
-        path: ranking_path.to_path_buf(),
-        metadata_path: None,
-        analysis_path: None,
-        analyzer_id: analyzer_id.to_string(),
-        source: RankingSource::LocalCache,
-        reason: "local analysis artifact".to_string(),
-    };
-    let model_files = discover_distribution_files(model)?;
-    let content =
-        serde_json::to_string_pretty(&build_analysis_json(model, &ranking, &model_files)?)? + "\n";
-    let analysis_path = ranking_path
-        .parent()
-        .unwrap_or_else(|| Path::new("."))
-        .join(ANALYSIS_JSON_FILENAME);
-    if let Some(parent) = analysis_path.parent() {
-        fs::create_dir_all(parent).with_context(|| format!("Create {}", parent.display()))?;
-    }
-    fs::write(&analysis_path, content)
-        .with_context(|| format!("Write {}", analysis_path.display()))?;
-    Ok(analysis_path)
-}
-
-pub(crate) fn write_package_analysis_json(
-    model: &MoeModelContext,
-    ranking: &ResolvedRanking,
-    output_path: &Path,
-) -> Result<()> {
-    let model_files = discover_distribution_files(model)?;
-    let content =
-        serde_json::to_string_pretty(&build_package_analysis_json(model, ranking, &model_files)?)?
-            + "\n";
-    if let Some(parent) = output_path.parent() {
-        fs::create_dir_all(parent).with_context(|| format!("Create {}", parent.display()))?;
-    }
-    fs::write(output_path, content).with_context(|| format!("Write {}", output_path.display()))
-}
-
-fn infer_analysis_details(
-    ranking: &ResolvedRanking,
-    model: &MoeModelContext,
-) -> Result<AnalysisDetails> {
-    let context_size_default = 4096;
-    let log_path = analysis_log_path(model, &ranking.analyzer_id);
-    let log_text = log_path
-        .as_ref()
-        .filter(|path| path.exists())
-        .map(fs::read_to_string)
-        .transpose()
-        .with_context(|| "Read local MoE analysis log".to_string())?;
-
-    let token_count = extract_first_arg_value(log_text.as_deref(), "-n")
-        .and_then(|value| value.parse::<u32>().ok())
-        .unwrap_or_else(|| {
-            if ranking.analyzer_id.starts_with("micro") {
-                128
-            } else {
-                32
-            }
-        });
-    let context_size = extract_first_arg_value(log_text.as_deref(), "-c")
-        .and_then(|value| value.parse::<u32>().ok())
-        .unwrap_or(context_size_default);
-    let all_layers = log_text
-        .as_deref()
-        .map(|text| text.contains("--all-layers"))
-        .unwrap_or(true);
-
-    if ranking.analyzer_id.starts_with("micro") {
-        let prompt_count = log_text
-            .as_deref()
-            .map(|text| text.matches("[prompt ").count())
-            .filter(|count| *count > 0)
-            .unwrap_or(8);
-        Ok((
-            Some("meshllm-micro-v1"),
-            Some(prompt_count),
-            token_count,
-            context_size,
-            all_layers,
-        ))
-    } else {
-        Ok((None, None, token_count, context_size, all_layers))
-    }
-}
-
-fn extract_first_arg_value<'a>(text: Option<&'a str>, flag: &str) -> Option<&'a str> {
-    let text = text?;
-    let parts = text.split_whitespace().collect::<Vec<_>>();
-    parts
-        .windows(2)
-        .find_map(|window| (window[0] == flag).then_some(window[1]))
-}
-
-pub(crate) fn analysis_log_path(model: &MoeModelContext, analyzer_id: &str) -> Option<PathBuf> {
-    let stem = model
-        .path
-        .file_stem()
-        .and_then(|value| value.to_str())
-        .unwrap_or("model");
-    Some(
-        crate::models::mesh_llm_cache_dir()
-            .join("moe")
-            .join("logs")
-            .join(format!("{stem}.{analyzer_id}.log")),
-    )
-}
-
-fn discover_distribution_files(model: &MoeModelContext) -> Result<Vec<(String, PathBuf)>> {
-    let Some(identity) = huggingface_identity_for_path(&model.path) else {
-        return Ok(vec![(
-            model
-                .path
-                .file_name()
-                .and_then(|value| value.to_str())
-                .unwrap_or("model.gguf")
-                .to_string(),
-            model.path.clone(),
-        )]);
-    };
-
-    let snapshot_root = snapshot_root_for_relative_file(&model.path, &identity.file)
-        .unwrap_or_else(|| {
-            model
-                .path
-                .parent()
-                .unwrap_or_else(|| Path::new("."))
-                .to_path_buf()
-        });
-    let mut files = Vec::new();
-    collect_distribution_files(
-        &snapshot_root,
-        &snapshot_root,
-        &model.distribution_id,
-        &mut files,
-    )?;
-    if files.is_empty() {
-        files.push((identity.file.clone(), model.path.clone()));
-    }
-    files.sort_by(|a, b| a.0.cmp(&b.0));
-    Ok(files)
-}
-
-fn snapshot_root_for_relative_file(path: &Path, relative_file: &str) -> Option<PathBuf> {
-    let mut root = path.to_path_buf();
-    for _ in Path::new(relative_file).components() {
-        root = root.parent()?.to_path_buf();
-    }
-    Some(root)
-}
-
-fn collect_distribution_files(
-    snapshot_root: &Path,
-    current: &Path,
-    distribution_id: &str,
-    files: &mut Vec<(String, PathBuf)>,
-) -> Result<()> {
-    for entry in fs::read_dir(current).with_context(|| format!("Read {}", current.display()))? {
-        let entry = entry?;
-        let path = entry.path();
-        if entry.file_type()?.is_dir() {
-            collect_distribution_files(snapshot_root, &path, distribution_id, files)?;
-            continue;
-        }
-        if path.extension().and_then(|ext| ext.to_str()) != Some("gguf") {
-            continue;
-        }
-        let relative = path
-            .strip_prefix(snapshot_root)
-            .unwrap_or(&path)
-            .to_string_lossy()
-            .replace('\\', "/");
-        if normalize_distribution_id(&relative) == distribution_id {
-            files.push((relative, path));
-        }
-    }
-    Ok(())
-}
-
-pub(crate) fn sha256_file(path: &Path) -> Result<String> {
-    let mut digest = Sha256::new();
-    let mut file = fs::File::open(path).with_context(|| format!("Open {}", path.display()))?;
-    let mut buf = [0u8; 1024 * 1024];
-    loop {
-        let read = std::io::Read::read(&mut file, &mut buf)
-            .with_context(|| format!("Hash {}", path.display()))?;
-        if read == 0 {
-            break;
-        }
-        digest.update(&buf[..read]);
-    }
-    Ok(format!("sha256:{:x}", digest.finalize()))
-}
-
-fn iso8601_now() -> String {
-    let now = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .unwrap_or_default()
-        .as_secs();
-    chrono::DateTime::<chrono::Utc>::from_timestamp(now as i64, 0)
-        .unwrap_or_else(chrono::Utc::now)
-        .to_rfc3339()
 }
 
 fn model_display_name(model_path: &Path) -> String {
